agents/deploy/requirements.txt
# Vertex AI Agent Engine dependencies
google-cloud-aiplatform>=1.112
google-adk>=1.0.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Numerical computing (for future strategy calculations)
numpy>=1.24.0

# HTTP requests (for API calls)
requests>=2.31.0
